refactor: remove cycle from Locations
