+++ src/components/Company/Locations/LocationForm/useLocationForm.ts
@@ -1,0 +1,12 @@
+import { createCompoundContext } from '@/components/Base'
+
+type LocationsFormContextType = {
+  isPending: boolean
+  handleCancel: () => void
+}
+
+const [useLocationsForm, LocationsFormProvider] = createCompoundContext<LocationsFormContextType>(
+  'CompanyDocumentFormContext',
+)
+
+export { useLocationsForm, LocationsFormProvider }
